クラスを分けてoption.rb作成、rubocop終了
