+++ 06.wc/wc.rb
@@ -4,13 +4,13 @@
 # frozen_string_literal: true
 
 require 'optparse'
+require_relative 'option'
 
 class Wc
   def initialize
-    @option_hash = { l: false, w: false, c: false, m: false }
+    @option = Option.new
     @files = []
     @answer_string = ''
-    @total_hash = { l: 0, w: 0, c: 0, m: 0 }
   end
 
   def wc(argv_array)
@@ -18,9 +18,11 @@ class Wc
     if argv_parse(argv_array)
       total = 0
 
-      # 標準入力か、パイプ
+      # コマンドライン引数にファイル名がない場合
       if @files.empty?
+        # 標準入力かパイプから読み込み、オプションを適用し答えの文字列を得る
         @answer_string += "\n" if count_file(nil)
+      # コマンドライン引数にファイル名がある場合
       else
         # ファイルごとに
         @files.each do |file|
@@ -29,20 +31,12 @@ class Wc
           total += 1
         end
       end
-      write_total if total > 1
+      # total の値を ＠answer_string に書き込む
+      @answer_string += @option.write_total if total > 1
     end
     # 答えの文字列を表示し返す
     print @answer_string
     @answer_string
-  end
-
-  # total の値を ＠answer_string に書き込む
-  def write_total
-    @answer_string += @total_hash[:l].to_s.rjust(8) if @option_hash[:l]
-    @answer_string += @total_hash[:w].to_s.rjust(8) if @option_hash[:w]
-    @answer_string += @total_hash[:c].to_s.rjust(8) if @option_hash[:c]
-    @answer_string += @total_hash[:m].to_s.rjust(8) if @option_hash[:m]
-    @answer_string += " total\n"
   end
 
   # コマンドライン引数を受け取り、オプションはハッシュに格納し、残りは配列に入れ空文字列を返す
@@ -50,15 +44,15 @@ class Wc
     # OptionParseオブジェクトoptを生成する
     opt = OptionParser.new
     # オプションを取り扱うブロックをoptに登録し、ハッシュに格納する
-    opt.on('-l') { @option_hash[:l] = true }
-    opt.on('-w') { @option_hash[:w] = true }
-    opt.on('-c') { @option_hash[:c] = true, @option_hash[:m] = false }
-    opt.on('-m') { @option_hash[:m] = true, @option_hash[:c] = false }
+    opt.on('-l') { @option.option_hash[:l] = true }
+    opt.on('-w') { @option.option_hash[:w] = true }
+    opt.on('-c') { @option.option_hash[:c] = true, @option.option_hash[:m] = false }
+    opt.on('-m') { @option.option_hash[:m] = true, @option.option_hash[:c] = false }
     # 残りのコマンドライン引数を配列に入れる
     begin
       @files = opt.order(argv_array)
       # オプションが指定されていない場合、デフォルトで行数、単語数、バイト数を表示する
-      @option_hash = { l: true, w: true, c: true } if @option_hash == { l: false, w: false, c: false, m: false }
+      @option.option_hash = { l: true, w: true, c: true } if @option.option_hash == { l: false, w: false, c: false, m: false }
     # 無効なオプションが指定されていた場合
     rescue OptionParser::ParseError => e
       @answer_string += "wc: illegal option -- #{e.args[0][1]}\nusage: wc [-clmw] [file ...]\n"
@@ -92,76 +86,9 @@ class Wc
     end
 
     # オプションを適用し、文字列を得る
-    apply_option(buf)
-  end
-
-  # オプションを適用し、文字列を得る
-  def apply_option(buf)
-    apply_l_option(buf) if @option_hash[:l]
-    apply_w_option(buf) if @option_hash[:w]
-    apply_c_option(buf) if @option_hash[:c]
-    apply_m_option(buf) if @option_hash[:m]
+    @answer_string += @option.apply_option(buf)
     true
-  end
-
-  # l オプションを適用し、1文字のスペースと７桁の数字からなる文字列を得る
-  def apply_l_option(buf)
-    # 行数を数える（@bufが不正なバイト列を含む場合?に置き換える）
-    n_line = buf.scrub.count("\n")
-    @answer_string += n_line.to_s.rjust(8)
-    @total_hash[:l] += n_line
-  end
-
-  # w オプションを適用し、1文字のスペースと７桁の数字からなる文字列を得る
-  def apply_w_option(buf)
-    # 文字列がascii文字だけの場合
-    if buf.ascii_only?
-      n_word = buf.split.count
-    else
-      # 空白文字
-      # 空白    ' ' : 0x20
-      # 改頁    \f  : 0x0C
-      # 改行    \n  : 0xA
-      # 復帰    \r  : 0x0D
-      # 水平タブ \t  : 0x09
-      # 垂直タブ \v  : 0x0B
-      # ノーブレークスペース  : 0x85
-      wspace_array = %w[20 0C A 0D 09 0B 85]
-
-      # 文字列をUTF-8で符号化しバイト単位に分割、16進数表記に変換する
-      buf_array = buf.bytes.map do |b|
-        b.to_s(16).upcase
-      end
-      # 空白文字で区切られた単語数を数える
-      word_flag = 1
-      n_word = 0
-      buf_array.each do |b|
-        if wspace_array.include?(b)
-          word_flag = 1
-        elsif word_flag == 1
-          word_flag = 0
-          n_word += 1
-        end
-      end
-    end
-    @answer_string += n_word.to_s.rjust(8)
-    @total_hash[:w] += n_word
-  end
-
-  # c オプションを適用し、1文字のスペースと７桁の数字からなる文字列を得る
-  def apply_c_option(buf)
-    # バイト数を数える（文字列のバイト長を整数で返す）
-    n_bytesize = buf.bytesize
-    @answer_string += n_bytesize.to_s.rjust(8)
-    @total_hash[:c] += n_bytesize
-  end
-
-  # m オプションを適用し、1文字のスペースと７桁の数字からなる文字列を得る
-  def apply_m_option(buf)
-    # 文字数を数える（マルチバイト文字1文字を1文字と数える）
-    n_size = buf.size
-    @answer_string += n_size.to_s.rjust(8)
-    @total_hash[:m] += n_size
+    # apply_option(buf)
   end
 end
